+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-nxt client
 unreleased
 ----------
 
+- implement temporal vector directive/layer
+
 - Rain bars are drawn and removed one by one
 
 - Converted land use donut to horizontal bar